+++ src/Components/Inputs/File.php
@@ -7,7 +7,7 @@ use Illuminate\Validation\Factory;
 use Northwestern\SysDev\DynamicForms\Components\BaseComponent;
 use Northwestern\SysDev\DynamicForms\Components\UploadInterface;
 use Northwestern\SysDev\DynamicForms\RuleBag;
-use Northwestern\SysDev\DynamicForms\Rules\FileExistsInS3;
+use Northwestern\SysDev\DynamicForms\Rules\FileExists;
 use Northwestern\SysDev\DynamicForms\Storage\StorageInterface;
 
 class File extends BaseComponent implements UploadInterface
@@ -23,7 +23,7 @@ class File extends BaseComponent implements UploadInterface
         } else {
             $rules->add('nullable');
         }
-        $rules->add(new FileExistsInS3($this->getStorageDriver()));
+        $rules->add(new FileExists($this->getStorageDriver()));
 
         return $validator->make(
             [$fieldKey => $submissionValue],
@@ -41,15 +41,19 @@ class File extends BaseComponent implements UploadInterface
         $this->storage = $storage;
     }
 
-    public function submissionValue(): mixed
+    /**
+     * Files always come in an array, even when in single-value mode.
+     */
+    protected function hasMultipleValuesForValidation(): bool
     {
-        if ($this->hasMultipleValues) {
-            return parent::submissionValue();
-        }
-        if ($this->submissionValue == []) {
-            return null;
-        }
+        return true;
+    }
 
-        return parent::submissionValue()[0] ?? parent::submissionValue();
+    /**
+     * @return string
+     */
+    public function getStorageType(): string
+    {
+        return $this->additional['storage'];
     }
 }