+++ src/DynamicFormsProvider.php
@@ -23,6 +23,10 @@ class DynamicFormsProvider extends ServiceProvider
         });
 
 
+        $this->app->singleton(FileComponentRegistry::class, function ($app) {
+            return new FileComponentRegistry;
+        });
+
         $this->app->singleton(S3Driver::class, function ($app) {
             $clientConfig = [
                 'region' => config('filesystems.disks.s3.region', Arr::get($_ENV, 'AWS_DEFAULT_REGION')),
@@ -59,6 +63,8 @@ class DynamicFormsProvider extends ServiceProvider
 
         /** @var JSONLogicInitHelper $jsonHelper */
         $jsonHelper = $this->app->make(JSONLogicInitHelper::class);
+        /** @var FileComponentRegistry $fileRegistry */
+        $fileRegistry = $this->app->make(FileComponentRegistry::class);
 
         Request::macro('validateDynamicForm', function (string $definitionJson, string $submissionJson) {
             $formDefinition = new Form($definitionJson); // @TODO pass registry